moveHorse function completed. Still can't retrieve horse from mouse click. Hell annoying

+++ src/model/Dice.java
@@ -15,6 +15,6 @@ public class Dice {
     }
 
     public void throwDice() {
-        diceValue = (int) (numberGenerator.nextInt(6) + 1);
+        diceValue = (numberGenerator.nextInt(6) + 1);
     }
 }
